change relative dir to warning and clean up code
